Fix a compilation error reported by llvm-gcc on x86-64: "input constraint with a matching output constraint of incompatible type"
git-svn-id: http://data-race-test.googlecode.com/svn/trunk@2631 93a3164d-3a44-0410-8ae5-4baa038b0729

+++ unittest/racecheck_unittest.cc
@@ -6553,7 +6553,7 @@ namespace  RepPrefixedInstructionsTest {  //{{{1
 #if defined (__GNUC__) && (defined(ARCH_x86) || defined(ARCH_amd64))
 void rep_clr_1(uint8_t *s, long n)
 {
-  int d0, d1;
+  intptr_t d0, d1;
       __asm__ __volatile__ (
       "rep ; stosb"
       : "=&c" (d0), "=&D" (d1)
